[Refactor] Move zerocoin checks out of main
This moves all zerocoin checking methods out of main.cpp/h into two new
files in the `consensus` subdir: zerocoin_verify.cpp/h
